Re-arrange TESTS dir to support unity-tests build system

+++ src/tinycbor.h
@@ -227,7 +227,7 @@ CBOR_API CborError cbor_encode_byte_string(CborEncoder *encoder, const uint8_t *
 CBOR_API CborError cbor_encode_floating_point(CborEncoder *encoder, CborType fpType, const void *value);
 
 CBOR_INLINE_API CborError cbor_encode_boolean(CborEncoder *encoder, bool value)
-{ return cbor_encode_simple_value(encoder, (int)value - 1 + (CborBooleanType & 0x1f)); }
+{ return cbor_encode_simple_value(encoder, (uint8_t)((int)value - 1 + (CborBooleanType & 0x1f))); }
 CBOR_INLINE_API CborError cbor_encode_null(CborEncoder *encoder)
 { return cbor_encode_simple_value(encoder, CborNullType & 0x1f); }
 CBOR_INLINE_API CborError cbor_encode_undefined(CborEncoder *encoder)
